Comments and Saving
I'm saving this version and creating another branch because I'm
unsure of how to implement the map and switching maps. Currently,
it's designed such that map.js handles all the logic for creating
an class. Then other js files will instantiate their own map.

However, switching between maps is proving to be difficult with this
design.