Map Switching and Ignore
The map switching now works. When a user exits the current map,
a new html file is loaded to the DOM. Each html will have its own
unique .js file where the next map will be created and init'd.

The js files will be stored in jsmaps and the html maps will be in html
maps.

Next, I will need to remember the locations that a player exits a gate,
so that when they return, they return through the correct gate.