Broken Gate Logic
To remember the last gate, I created a new process. When a new map is initialized, map.js
asks main.js, who asks bgProcess.js for the last map that we were in. Also, we have to remember
the last map when we change maps. The ipc communication is currently in place, but we
have to do some serious debugging.

The bgProcess will always be running. I'm sure I'll think of other functions and processes
that need to go there eventually.

ipc Inter process communication. In the renderer files, the ipcRenderer needs to be
called. In the main process, the ipcMain needs to be required. I can't communicate
from renderer to renderer, but I can go from renderer to main, then main to another
renderer.