Created Random Moving NPC
Created the logic for a moving npc. The function to create an npc is called
insert person and can be found in the Map.exports. The calculation for the
direction of each movement is in bgProcess.js.

An npc will randomly move every 3 seconds in a valid direction.

Known bugs
Can't have more than one npc
Memory leak after about 10 calls from event emitter