Functionality for Multiple NPC
Every map has the ability to randomly generate a crowd. NPCs of the crowd will
walk around the map, but for some reason, after a long period of time, the
crowd will move only to the far left side of the map, then wander up
and down the border.

Known bugs
- memory leaks after a long time
- Math.random() stops generating actual randos